pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]

[project]
name = "asyncio_play"
version = "0.2.0"
description = "A project to play with asyncio"
authors = [
    {name="Michael Penhallegon", email="mike@hematite.tech"}
]
dependencies = [
    "grpcio",
    "grpcio-tools"
]

[project.optional-dependencies]
test = [
    "pytest",
    "pytest-asyncio",
    "tox"
]
debug = [
    "ipdb"
]
linting = [
    "isort",
    "black",
    "ruff"
]

[options]
python_requires = ">=3.8"
install_requires = ["wheel", "asyncio >= 3.4"]

[options.package.find]
where=["src"]
exclude=["script", "protos"]

[tool.ruff]
line-length = 120
exclude = ["protos/**.py", "src/activitypub_service/*.py"]

[tools.pytest.ini_options]
minversion = "6.0"
markers = [
    "integration: mark a test as an integration test",
]